+++ app/charts/column/columns-state.tsx
@@ -53,7 +53,11 @@ import {
   formatNumberWithUnit,
   useTimeFormatUnit,
 } from "@/formatters";
-import { TemporalDimension, TimeUnit } from "@/graphql/query-hooks";
+import {
+  DimensionMetadataFragment,
+  TemporalDimension,
+  TimeUnit,
+} from "@/graphql/query-hooks";
 import { getPalette } from "@/palettes";
 import {
   getSortingOrders,
@@ -79,7 +83,7 @@ export interface ColumnsState extends CommonChartState {
   segments: string[];
   colors: ScaleOrdinal<string, string>;
   yAxisLabel: string;
-  yAxisDescription: string | undefined;
+  yAxisDimension: DimensionMetadataFragment;
   getAnnotationInfo: (d: Observation) => TooltipInfo;
   showStandardError: boolean;
 }
@@ -229,7 +233,6 @@ const useColumnsState = (
   }
 
   const yAxisLabel = getLabelWithUnit(yMeasure);
-  const yAxisDescription = yMeasure.description || undefined;
 
   const { left, bottom } = useChartPadding(
     yScale,
@@ -358,7 +361,7 @@ const useColumnsState = (
     getSegment,
     getSegmentLabel,
     yAxisLabel,
-    yAxisDescription,
+    yAxisDimension: yMeasure,
     segments,
     colors,
     getAnnotationInfo,