+++ app/components/chart-published.tsx
@@ -3,6 +3,7 @@ import { Box, Theme, Typography } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import * as React from "react";
 import { useEffect, useMemo, useRef } from "react";
+import { useStore } from "zustand";
 
 import { ChartDataFilters } from "@/charts/shared/chart-data-filters";
 import { isUsingImputation } from "@/charts/shared/imputation";
@@ -19,7 +20,11 @@ import {
 import GenericChart from "@/components/common-chart";
 import Flex from "@/components/flex";
 import { HintBlue, HintRed, HintYellow } from "@/components/hint";
-import { MetadataPanel } from "@/components/metadata-panel";
+import {
+  MetadataPanel,
+  MetadataPanelStoreContext,
+  createMetadataPanelStore,
+} from "@/components/metadata-panel";
 import {
   ChartConfig,
   ConfiguratorStatePublishing,
@@ -28,7 +33,7 @@ import {
   PublishedConfiguratorStateProvider,
 } from "@/configurator";
 import { DataSetTable } from "@/configurator/components/datatable";
-import { flag } from "@/configurator/components/flag";
+import { DRAWER_WIDTH } from "@/configurator/components/drawer";
 import { parseDate } from "@/configurator/components/ui-helpers";
 import {
   DEFAULT_DATA_SOURCE,
@@ -68,15 +73,18 @@ export const ChartPublished = ({
   );
 };
 
-const useStyles = makeStyles((theme: Theme) => ({
+const useStyles = makeStyles<Theme, { shrink: boolean }>((theme) => ({
   root: {
     position: "relative",
     display: "flex",
     flexGrow: 1,
     flexDirection: "column",
     padding: theme.spacing(5),
+    paddingLeft: ({ shrink }) =>
+      `calc(${theme.spacing(5)} + ${shrink ? DRAWER_WIDTH : 0}px)`,
     color: theme.palette.grey[800],
     overflowX: "auto",
+    transition: "padding 0.25s ease",
   },
 }));
 
@@ -93,11 +101,42 @@ export const ChartPublishedInner = ({
   chartConfig: ChartConfig;
   configKey: string;
 }) => {
-  const classes = useStyles();
+  const rootRef = useRef<HTMLDivElement>(null);
+
+  const {
+    state: isTablePreview,
+    setState: setIsTablePreview,
+    containerRef,
+    containerHeight,
+    computeContainerHeight,
+  } = useChartTablePreview();
+
+  const metadataPanelStore = useMemo(() => createMetadataPanelStore(), []);
+  const metadataPanelOpen = useStore(metadataPanelStore, (state) => state.open);
+
+  const shouldShrink = useMemo(() => {
+    const rootWidth = rootRef.current?.getBoundingClientRect().width;
+
+    if (!rootWidth) {
+      return false;
+    }
+
+    return metadataPanelOpen && rootWidth > DRAWER_WIDTH * 2;
+  }, [metadataPanelOpen]);
+
+  useEffect(() => {
+    const unsubscribe = metadataPanelStore.subscribe(() => {
+      computeContainerHeight();
+    });
+
+    return () => unsubscribe();
+  });
+
+  const classes = useStyles({
+    shrink: shouldShrink,
+  });
   const locale = useLocale();
   const isTrustedDataSource = useIsTrustedDataSource(dataSource);
-
-  const rootRef = useRef<HTMLDivElement>(null);
 
   const [{ data: metaData }] = useDataCubeMetadataWithComponentValuesQuery({
     variables: {
@@ -115,13 +154,6 @@ export const ChartPublishedInner = ({
       chartConfig: chartConfig,
     } as ConfiguratorStatePublishing;
   }, [chartConfig, dataSource]);
-
-  const {
-    state: isTablePreview,
-    setState: setIsTablePreview,
-    containerRef,
-    containerHeight,
-  } = useChartTablePreview();
   const handleToggleTableView = useEvent(() => setIsTablePreview((c) => !c));
 
   const allDimensions: DimensionMetadataFragment[] = useMemo(() => {
@@ -132,98 +164,96 @@ export const ChartPublishedInner = ({
   }, [metaData?.dataCubeByIri?.dimensions, metaData?.dataCubeByIri?.measures]);
 
   return (
-    <Box className={classes.root} ref={rootRef}>
-      <ChartErrorBoundary resetKeys={[chartConfig]}>
-        {metaData?.dataCubeByIri?.publicationStatus ===
-          DataCubePublicationStatus.Draft && (
-          <Box sx={{ mb: 4 }}>
-            <HintRed iconName="datasetError" iconSize={64}>
-              <Trans id="dataset.publicationStatus.draft.warning">
-                Careful, this dataset is only a draft.
-                <br />
-                <strong>Don&apos;t use for reporting!</strong>
-              </Trans>
-            </HintRed>
-          </Box>
-        )}
-        {metaData?.dataCubeByIri?.expires && (
-          <Box sx={{ mb: 4 }}>
-            <HintRed iconName="datasetError" iconSize={64}>
-              <Trans id="dataset.publicationStatus.expires.warning">
-                Careful, the data for this chart has expired.
-                <br />
-                <strong>Don&apos;t use for reporting!</strong>
-              </Trans>
-            </HintRed>
-          </Box>
-        )}
-        {!isTrustedDataSource && (
-          <Box sx={{ mb: 4 }}>
-            <HintYellow iconName="hintWarning">
-              <Trans id="data.source.notTrusted">
-                This chart is not using a trusted data source.
-              </Trans>
-            </HintYellow>
-          </Box>
-        )}
-        {isUsingImputation(chartConfig) && (
-          <Box sx={{ mb: 4 }}>
-            <HintBlue iconName="hintWarning">
-              <Trans id="dataset.hasImputedValues">
-                Some data in this dataset is missing and has been interpolated
-                to fill the gaps.
-              </Trans>
-            </HintBlue>
-          </Box>
-        )}
-        <Flex sx={{ justifyContent: "space-between", alignItems: "center" }}>
-          <Typography component="div" variant="h2" mb={2}>
-            {meta.title[locale]}
-          </Typography>
+    <MetadataPanelStoreContext.Provider value={metadataPanelStore}>
+      <Box className={classes.root} ref={rootRef}>
+        <ChartErrorBoundary resetKeys={[chartConfig]}>
+          {metaData?.dataCubeByIri?.publicationStatus ===
+            DataCubePublicationStatus.Draft && (
+            <Box sx={{ mb: 4 }}>
+              <HintRed iconName="datasetError" iconSize={64}>
+                <Trans id="dataset.publicationStatus.draft.warning">
+                  Careful, this dataset is only a draft.
+                  <br />
+                  <strong>Don&apos;t use for reporting!</strong>
+                </Trans>
+              </HintRed>
+            </Box>
+          )}
+          {metaData?.dataCubeByIri?.expires && (
+            <Box sx={{ mb: 4 }}>
+              <HintRed iconName="datasetError" iconSize={64}>
+                <Trans id="dataset.publicationStatus.expires.warning">
+                  Careful, the data for this chart has expired.
+                  <br />
+                  <strong>Don&apos;t use for reporting!</strong>
+                </Trans>
+              </HintRed>
+            </Box>
+          )}
+          {!isTrustedDataSource && (
+            <Box sx={{ mb: 4 }}>
+              <HintYellow iconName="hintWarning">
+                <Trans id="data.source.notTrusted">
+                  This chart is not using a trusted data source.
+                </Trans>
+              </HintYellow>
+            </Box>
+          )}
+          {isUsingImputation(chartConfig) && (
+            <Box sx={{ mb: 4 }}>
+              <HintBlue iconName="hintWarning">
+                <Trans id="dataset.hasImputedValues">
+                  Some data in this dataset is missing and has been interpolated
+                  to fill the gaps.
+                </Trans>
+              </HintBlue>
+            </Box>
+          )}
+          <Flex sx={{ justifyContent: "space-between", alignItems: "center" }}>
+            <Typography component="div" variant="h2" mb={2}>
+              {meta.title[locale]}
+            </Typography>
 
-          {flag("metadata") && (
             <MetadataPanel
               datasetIri={dataSet}
               dataSource={dataSource}
               dimensions={allDimensions}
               container={rootRef.current}
             />
-          )}
-        </Flex>
-
-        {meta.description[locale] && (
-          <Typography component="div" variant="body1" mb={2}>
-            {meta.description[locale]}
-          </Typography>
-        )}
-        <InteractiveFiltersProvider>
-          <Flex
-            flexDirection="column"
-            ref={containerRef}
-            height={containerHeight.current!}
-            flexGrow={1}
-          >
-            <PublishedConfiguratorStateProvider
-              chartId={configKey}
-              initialState={publishedConfiguratorState}
-            >
-              {isTablePreview ? (
-                <DataSetTable
-                  sx={{ maxHeight: "100%" }}
-                  dataSetIri={dataSet}
-                  dataSource={dataSource}
-                  chartConfig={chartConfig}
-                />
-              ) : (
-                <ChartWithInteractiveFilters
-                  dataSet={dataSet}
-                  dataSource={dataSource}
-                  chartConfig={chartConfig}
-                />
-              )}
-            </PublishedConfiguratorStateProvider>
           </Flex>
-          {chartConfig && (
+
+          {meta.description[locale] && (
+            <Typography component="div" variant="body1" mb={2}>
+              {meta.description[locale]}
+            </Typography>
+          )}
+          <InteractiveFiltersProvider>
+            <Flex
+              flexDirection="column"
+              ref={containerRef}
+              height={containerHeight.current!}
+              flexGrow={1}
+            >
+              <PublishedConfiguratorStateProvider
+                chartId={configKey}
+                initialState={publishedConfiguratorState}
+              >
+                {isTablePreview ? (
+                  <DataSetTable
+                    sx={{ maxHeight: "100%" }}
+                    dataSetIri={dataSet}
+                    dataSource={dataSource}
+                    chartConfig={chartConfig}
+                  />
+                ) : (
+                  <ChartWithInteractiveFilters
+                    dataSet={dataSet}
+                    dataSource={dataSource}
+                    chartConfig={chartConfig}
+                  />
+                )}
+              </PublishedConfiguratorStateProvider>
+            </Flex>
             <ChartFootnotes
               dataSetIri={dataSet}
               dataSource={dataSource}
@@ -231,10 +261,10 @@ export const ChartPublishedInner = ({
               configKey={configKey}
               onToggleTableView={handleToggleTableView}
             />
-          )}
-        </InteractiveFiltersProvider>
-      </ChartErrorBoundary>
-    </Box>
+          </InteractiveFiltersProvider>
+        </ChartErrorBoundary>
+      </Box>
+    </MetadataPanelStoreContext.Provider>
   );
 };