+++ app/components/select-tree.tsx
@@ -265,8 +265,7 @@ export type SelectTreeProps = {
   controls?: React.ReactNode;
   onChange: (ev: { target: { value: NodeId } }) => void;
   disabled?: boolean;
-  tooltipText?: string;
-  label?: string;
+  label?: React.ReactNode;
   onOpen?: () => void;
   onClose?: () => void;
   open?: boolean;
@@ -287,7 +286,6 @@ function SelectTree({
   options,
   value,
   onChange,
-  tooltipText,
   disabled,
   controls,
   onOpen,
@@ -478,7 +476,7 @@ function SelectTree({
   return (
     <div>
       {label && (
-        <Label htmlFor={id!} smaller tooltipText={tooltipText} sx={{ mb: 1 }}>
+        <Label htmlFor={id!} smaller sx={{ mb: 1 }}>
           {label} {controls}
         </Label>
       )}