+++ app/configurator/components/field.tsx
@@ -148,18 +148,16 @@ export const DataFilterSelect = ({
   isOptional,
   controls,
   hierarchy,
-  tooltipText,
   onOpen,
   loading,
 }: {
   dimension: DimensionMetadataFragment;
-  label: string;
+  label: React.ReactNode;
   id: string;
   disabled?: boolean;
   isOptional?: boolean;
   controls?: React.ReactNode;
   hierarchy?: HierarchyValue[];
-  tooltipText?: string;
   onOpen?: () => void;
   loading?: boolean;
 }) => {
@@ -222,7 +220,6 @@ export const DataFilterSelect = ({
         open={isOpen}
         disabled={disabled}
         controls={controls}
-        tooltipText={tooltipText}
         {...fieldProps}
       />
     );
@@ -236,7 +233,6 @@ export const DataFilterSelect = ({
       options={allValues}
       sortOptions={false}
       controls={controls}
-      tooltipText={tooltipText}
       open={isOpen}
       onClose={handleClose}
       onOpen={handleOpen}
@@ -257,7 +253,7 @@ export const DataFilterSelectDay = ({
   controls,
 }: {
   dimension: DimensionMetadataFragment;
-  label: string;
+  label: React.ReactNode;
   disabled?: boolean;
   isOptional?: boolean;
   controls?: React.ReactNode;
@@ -349,10 +345,9 @@ export const DataFilterSelectTime = ({
   disabled,
   isOptional,
   controls,
-  tooltipText,
 }: {
   dimension: DimensionMetadataFragment;
-  label: string;
+  label: React.ReactNode;
   from: string;
   to: string;
   timeUnit: TimeUnit;
@@ -361,7 +356,6 @@ export const DataFilterSelectTime = ({
   disabled?: boolean;
   isOptional?: boolean;
   controls?: React.ReactNode;
-  tooltipText?: string;
 }) => {
   const fieldProps = useSingleFilterSelect({ dimensionIri: dimension.iri });
   const formatLocale = useTimeFormatLocale();
@@ -416,7 +410,6 @@ export const DataFilterSelectTime = ({
         options={allOptions}
         sortOptions={false}
         controls={controls}
-        tooltipText={tooltipText}
         {...fieldProps}
       />
     );
@@ -430,7 +423,6 @@ export const DataFilterSelectTime = ({
       timeFormat={timeFormat}
       formatLocale={formatLocale}
       isOptional={isOptional}
-      tooltipText={tooltipText}
       onChange={fieldProps.onChange}
     />
   );
@@ -443,16 +435,14 @@ export const TimeInput = ({
   timeFormat,
   formatLocale,
   isOptional,
-  tooltipText,
   onChange,
 }: {
   id: string;
-  label: string;
+  label: React.ReactNode;
   value: string | undefined;
   timeFormat: string;
   formatLocale: TimeLocaleObject;
   isOptional: boolean | undefined;
-  tooltipText?: string;
   onChange: (e: ChangeEvent<HTMLInputElement>) => void;
 }) => {
   const [inputValue, setInputValue] = useState(
@@ -492,7 +482,6 @@ export const TimeInput = ({
       name={id}
       label={label}
       value={inputValue}
-      tooltipText={tooltipText}
       onChange={onInputChange}
     />
   );
@@ -541,14 +530,12 @@ export const MetaInputField = ({
   locale,
   value,
   disabled,
-  tooltipText,
 }: {
   label: string | ReactNode;
   metaKey: string;
   locale: string;
   value?: string;
   disabled?: boolean;
-  tooltipText?: string;
 }) => {
   const field = useMetaField({
     metaKey,
@@ -556,14 +543,7 @@ export const MetaInputField = ({
     value,
   });
 
-  return (
-    <Input
-      label={label}
-      {...field}
-      disabled={disabled}
-      tooltipText={tooltipText}
-    />
-  );
+  return <Input label={label} {...field} disabled={disabled} />;
 };
 
 const useMultiFilterColorPicker = (value: string) => {